+++ src/engine/Util/Required.ts
@@ -1,0 +1,3 @@
+export type DeepRequired<T> = Required<{
+  [K in keyof T]: T[K] extends Required<T[K]> ? T[K] : DeepRequired<T[K]>
+}>